requirements.txt
torch==1.11.0
torchvision==0.12.0
numpy==1.24.3
tqdm
tensorboard
matplotlib
pandas

axel
datasets
gdown
inflect
wandb
wget
torchaudio~=0.11.0
pillow
pyworld
unidecode
